test/requirements.txt
torch>=1.11.0
numpy>=1.21.0
scikit-learn>=1.0.0
scapy>=2.4.5
mininet>=2.3.0.dev6
PyYAML>=6.0
watchdog>=2.1.0
psutil>=5.8.0
colorlog>=6.6.0
tensorboard>=2.8.0
ryu>=4.34
eventlet>=0.33.0